tools/backup-viewer/requirements.txt
flask==3.1.1
Pillow>=10.0.0
